// src/features/billing/enums/enums.ts
export enum CalculationType {
  FIXED = 'FIXED',
  PERCENTAGE = 'PERCENTAGE',
}

export enum ServiceType {
  TOUR = 'TOUR',
  HOTEL = 'HOTEL',
}

export enum PlaceType {
  HOTEL = 'HOTEL',
  HOUSEHOLDER = 'HOUSEHOLDER',
  HOSTEL = 'HOSTEL',
  GUEST_HOUSE = 'GUEST_HOUSE',
  APARTMENT = 'APARTMENT',
  PRIVATE_HOUSE = 'PRIVATE_HOUSE',
  RESORT = 'RESORT',
  SANATORIUM = 'SANATORIUM',
  CAMPING = 'CAMPING',
  ALL = 'ALL',
}

export enum ProviderType {
  LOCAL = 'LOCAL',
  INTERNATIONAL = 'INTERNATIONAL',
}

export enum CitizenType {
  LOCAL = 'LOCAL',
  FOREIGN = 'FOREIGN',
}

export enum ProviderCode {
  CLICK = 'CLICK',
}